third_party/blink/web_tests/fast/canvas/canvas-text-drawing-style-letterWordSpacing.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<body>
<script>
test(function(t) {
    var ctx = document.createElement('canvas').getContext('2d');

    // Verify initial values:
    assert_equals(ctx.textLetterSpacing, 0);
    assert_equals(ctx.wordSpacing, 0);

    ctx.textLetterSpacing = 3;
    assert_equals(ctx.textLetterSpacing, 3);
    assert_equals(ctx.wordSpacing, 0);
    ctx.wordSpacing = 5;
    assert_equals(ctx.textLetterSpacing, 3);
    assert_equals(ctx.wordSpacing, 5);

}, "Test basic functionality of textLetterSpacing and wordSpacing.");
</script>
</body>
